refactor: separate server code
